MXS-2481 Add MaxCtrl class to test framework
The purpose of this class is to make it easy from a test program
to access the output of the REST-API. Currently it provides only
the equivalent of "maxctrl list servers", but it can be extended
if and when additional needs arise.

Right now it's implemented so that curl is executed on the 0th
MaxScale instance, using the ssh mechanism of the test framework.