Duplicate RWBackend request tracking in MySQLProtocol
This is the initial step towards the removal of the duplication of request
state tracking when multiple modules need to track the state of a
response. The next step is to encapsulate the request state as an object
and pass it along to the clientReply method.

This does not yet fully fulfill the needs of all filters: the masking
filter requires more detailed information which needs to be incorporated
into the new code.