Reorganize dcb_connect
Moved the persistent connection creation into a separate function to make
the connection logic easier to comprehend. Also separated and reordered
object allocation so that a the smallest amount of deallocations would
have to be done if one allocation fails. Splitting the allocation of the
DCB and the actual connection into separate parts made this easy and
further clarifies the separation of "important work" from the basic DCB
setup.

Removed some debug messages as they were either redundant or were
implicitly expressed by errors or a lack of messages.

A change was also made that forbids protocol modules without the
auth_default entry point. Currently all modules define it and if it wasn't
defined, be it due to either by a bug or a new protocol module, unexpected
behavior would most likely occur. This would also imply that the NullAuth
authenticator should be removed from MaxScale as they are unlikely to work
at all (see https://jira.mariadb.org/browse/MXS-2499 for a case where
mixed authenticators cause problems).