Make session state and ID private
The session state should not be modified outside of the internal
functions.

Renamed the SESSION_STATE_TO_BE_FREED to better represent its actual use
as a signal that the session creation failed.

Moved the various session startup and shutdown functions inside the
Session class so that they are not exposed to the outside world.