Remove null auth modules
The null_auth_allow module wouldn't work as it didn't store the password
hash that's required for backend connections. The skip_authentication
option in the MariaDB protocols can be used to replace this.

The null_auth_deny function is mostly useless in practice and is thus only
a maintenance burden.