Make Session a mxs::Component
By making the session a mxs::Component it can be used as the upstream
component of a service or a server endpoint. With this change, the target
method no longer makes as much sense in the Component class as it would in
the Endpoint class.

As the ServiceEndpoint now owns the router sessions, the router session in
the MXS_SESSION and all the related code can be removed.