Fix HTTP protocol authentication
The HTTP protocol used the NullAuthAllow authenticator to skip the
authentication. There was also a httpauth module that was added at some
point but was not documented or used by any module. Repurposed the
httpauth module to be a substitute for the now removed NullAuthAllow
authenticator and changed the default plugin in the HTTP protocol module.