Add Target::get_children
The method can be used to get the child nodes of a routing target. This
makes it possible for services to get the list of leaf nodes in the
routing tree which in turn can be used to retrieve the list of servers
that a service can reach.

The list of targets is stored in the SERVICE::Config struct but it does
not belong there. A later commit will move it as an internal member of
Service.