Use SERVICE::servers to get a list of reachable servers
The method provides a list of servers that the authenticators can use to
extract database user information. The function implementation is done in
a separate commit as it involves moving member variables around into a
worker-local storage internal to the Service.

The n_dbref was also removed as this is already provided by the
Target::get_children::size method.