Keep service internal data inside a rworker_local container
By storing the internal service data inside a member variable that's local
to the worker, we make all access into it thread-safe. This removes the
need for the `dbref` linked list and simplifies iteration operations by
storing everything in a STL container.

The values for the list of reachable servers and the minimum and maximum
server version values are pre-calculated every time a new target is added
to the service. This reduces the cost of reading them but increases the
cost of writing but given the relative rarity of modifications, this is an
acceptable cost.