Simplify the use of the handleError API method
By using the clientReply function to propagate errors to the client, the
handleError method is converted into a single-purpose API function which
only handles backend connection failures. This, in turn, reduces the
amount of code that the routers need to implement as well as the overall
complexity of the functions.

The function signature now also uses a return value to signal whether the
session can continue. This is an improvement as it removes the ambiguity
of not returning a value from the function.