MXS-1386 Remove owning_dcb from MySQLProtocol
The precense of 'owning_dcb' in MySQLProtocol creates a circular
dependency between MySQLProtocol and DCB.
- When a DCB is created, its m_protocol needs to be assigned.
- m_protocol will be a MySQLProtocol that cannot be created without
  a DCB.
Thus, the creation of a DCB must be two-phase; first the DCB is
created without assigning m_protocol, then the MySQLProtocol is
indirectly created, and finally m_protocol is assigned.

This is the first step in breaking that dependency, so that all
data needed by a DCB can be created before the actual DCB is created.