Use mxs::Endpoint in readwritesplit
A large part of the changes were about changing from explicit SERVER usage
to mxs::Target usage via mxs::Endpoint.

Removed the client DCB pointer from the RWSplitSession. This also fixes
the problem where generated responses would not go through the filter
chain.

Removed router connection counts and server weights as they are no longer
available when the new system is in use.

Some parts of readwritesplit are not functional due to the fact that the
last write and read variables that the DCBs used to expose are no longer
available. This needs to be added somewhere, possibly inside RWBackend or
mxs::Component, where the routers can safely access it.