Use mxs::Endpoint in schemarouter
The changes consist mainly of renames from SERVER to mxs::Target and
SERVER_REF to mxs::Endpoint.

The SRBackend objects are now also stored in unique pointers instead of
shared pointers. This retains the automatic lifetime management but
removes the unnecessary capability to share ownership of the pointers.

All log messages that referred to hosts and ports have been changed to use
the unique object names instead.