Fix connection_keepalive in readwritesplit
Moved the last write implementation into the rwbackend to remove the need
to access the DCB internals. Replace use of internal MaxScale clock with
chrono and use chrono types to make time comparisons easier. Remove use of
DCBs from the modutil ping function and convert it to a function that
creates a GWBUF.