Move variables from SERVICE to Service
The parameters are now a part of the internal Service class and a method
is provided that allows read-only access to it. The modifications to the
parameters also now go through the service class which removes the need
for a few C-style functions.

Also changed the service state from a set of defines into an enum class
and provided default values in the public member variable declarations.