Add back rlag log messages
Moved the rlag state handling into Target so that it works the same way
for both services and servers.

The assignment is done using compare-and-swap which prevents most
duplicate log messages but as there's no unique identifier for each state
changes, it does not guarantee absolute ordering of state changes. Since
it's only a log message it's not a problem.