Fix query retrying in readwritesplit
The query retrying used a reinterpret cast from RWSplitSession to
MXS_FILTER_SESSION when it first should be static cast into a
MXS_ROUTER_SESSION. The reinterprent cast would store the pointer to the
derived class which would not convert back to the original value when the
router template's routeQuery function converts the base class pointer to
the derived class pointer.

RWSplitSession should also be the final class and nothing should be
inherited from it. This allows the removal of the empty virtual
destructor.