Make mxs::Reply in clientReply a const reference
This makes the use of the reply simpler by removing the possibility of
null replies. All replies that are given to clientReply should also have
an accompanying reply object. In most cases this is the one that is stored
in the protocol module but for router-generated responses a separate reply
object must be constructed.