Fix smartrouter KILL generation
The KILL generation now uses the session ID to send the kill to all
connections created for that session. This does cause one extra KILL to be
sent. The server where the fastest reply was returned will also receive
this but this is assumed to not be a big performance problem.