Call module init functions after startup
If a module was loaded after the workers have started, the process and
thread init functions would not be called. This can be fixed by
broadcasting a message to all workers to execute the thread init function.

In order for us to know if these functions need to be called, the
RoutingWorker class must expose whether the worker threads have been
started.