Fix LocalClient buffer management
The queries passed to queue_query are now assumed to be owned by
LocalClient, the same way the routeQuery functions behave.

The freeing of the buffer in LocalClient::handleError had to be removed as
it caused a double free due to the fact that the buffer ownership is not
passed along to the handleError function. This behavior should be reviewed
and possibly made consistent with the other API functions by passing
ownership of the buffer.