MXS-1386 Drop dcb_session_check()
Replaced with direct calls to DCB::ready().

I think it should not be possible that an event is delivered to
a DCB that is not ready. A DCB should also not be able to turn
non-ready during the processing of events. For the time being
that's still checked using both an assert and an if.