Fix KILL query handling
The information collection is now offloaded to a separate thread to
prevent deadlocks. As the LocalClient can now only be used on the thread
that creates the session, the actual execution of the KILL query is done
the original thread once the information about the connections that need
to be killed is collected.

The session_get_ref function is now exposed to the modules to prevent the
session from being closed before the LocalClient instances have been
created. Some internal mechanism could be created that allows safe
execution of code even after the session has been closed.

The KILL execution does have one remaining problem that needs to be
addressed: there can be multiple connections to the same server. This
should be rather easy to solve by storing all queries instead of only the
latest one.