MXS-1386 Introduce DCB::m_handler
The epoll_wait notifications are now delivered in terms of
DCB::Handler and not MXS_PROTOCOL_SESSION. In practice the
handler is always either a client or backend protocol object,
currently stored in ClientDCB and BackendDCB respectively.

The need for DCB being aware of its protocol object disappears
once the DCB become the property of the protocol object and
invisible to everyone else. At that point DCB becomes a proper
abstraction for asynchronous socket communication and can be
moved to maxutils.