Use actual object names in log messages
Due to the recursive nature of service-to-service routing, adding the
object names to the log messages greatly improves their readability.

This commit adds the LogScope class and takes it into use ServiceEndpoint
and ServerEndpoint. Messages logged by filters and routers now contain the
object name instead of the module name.