Fix slave connection counts in readwritesplit
If no explicit limits have been configured, the default should be to allow
as many slave connections as there are configured servers in the
service. If all the servers configured for the service would be slaves,
all but one of them would be used even if no limit was configured.