MXS-2656 Send COM_CHANGE_USER in reuse_connection()
COM_CHANGE_USER is now sent in reuse_connection(). That means that
there no longer is a need for the protocol to know in its write()
function whether the DCB used to be persistent (and indeed it no
longer will be). That is, m_was_persistent can now be removed.