Fix protocol level resultset collection
The functionality is used by routers to collect various responses into one
buffer. Since the result was stored in the DCB's readq until it was
complete, it was possible that a single result would be read "multiple"
times which in turn greatly confused the request tracking code. To cope
with this, the results are now stored in a separate buffer if request
tracking is enabled.