MXS-2617 Replace spinlock with a mutex and condition variables
Capping correctly waited for pointer releases, but forgot to call gc().
Also realized that the already calculated gc_ptr_count can be used
instead of calculating in_use_ptrs.

Waiting for garbage is a timeout and re-check, rather than waiting
on the condition variable. The reason is that the cost of locking and
notifying in reader_ready() is pretty high, the test workers became
8-10% slower.