Make rworker_local::assign synchronous
By forcing the assign method to wait on all the workers to apply the new
value, we can be certain that the state of the stored object is the same
on all workers once the function returns. In practice this allows
modifications via rworker_local::apply to be done as if the value was a
normal member variable. The only downside is that the update must be done
on a specific thread which is already handled by the current code.

One case where this simplifies things is when the targets of a service are
being modified before its destruction. Once all of the targets are
removed, we can be certain that no new sessions will be created for that
service.

Also changed rworker_local::values to use execute_concurrently.