Fix mxs1731_old_persisted_config
The service isn't serialized if the parameter value doesn't change when it
is modified via the REST API. The old version which used maxadmin always
caused the configuration to be serialized even if no changes were done.

By setting enable_root_user to true then false we'll be sure to change it
at least one.