MXS-2704 Introduce maxscale::MaxScaleWorker
MaxScaleWorker is a worker class from which all MaxScale worker
classes must be derived, with the exception of MainWorker.
MaxScaleWorker will provide the information needed for MainWorker
to be able to notify the systemd watchdog in a timely manner. With
this change, the special role that the main routing worker sofar
has had with respect to systemd watchdog notifications will be
removed.