MXS-2009: Asychronous session command execution
With the addition of the session command replier pointer, the execution of
session commands can be made asynchronous. In practice this means that new
session commands are queued on all servers but only the response of the
pre-assigned replier is waited for.

This should improve the overall responsiveness of readwritesplit as well
as greatly reduce the worst-case latency for session commands. Slower
slave servers no longer cause the whole session to stall.