MXS-2009: Fix session command retrying
The session command retrying that happens when a master fails didn't work
due to the way successful session command executions were counted. In
addition to this, it was possible that a backend had two session commands
with the same ID in the queue.

By electing one of the surviving servers as the session command replier,
the need to re-execute the session command goes away in some cases. The
only case when it needs to be re-executed is when all slaves have already
responded before the master fails.