Add states for cluster operations
The operations that modify the cluster now update a state variable which
is displayed in the monitor output. This allows the operation state to be
monitored via the REST API when it is done automatically by the
monitor. For manual operations the information is useless as the REST API
is blocked during the operation.