MXS-760: Make authentication failures normal errors
The authentication failure is now something that the router can
handle. Although it's not possible to successfully create a connection to
the server where the authentication fails, giving the control over to the
router allows the router session to continue even if authentication
failures happen.

Currently the information about the authentication failure is implicitly
present in the error that is returned to the client. If the error buffer
contains a ER_ACCESS_DENIED_ERROR, the authentication has failed and the
server should not be used again. Making this information explicit by
passing it as an argument to the handleError entry point might be a better
way to convey the information.