MXS-760: Pass error type to handleError
By telling the upstream Component what sort of an error happened, it can
act accordingly. Transient errors should be handled by reconnecting to the
endpoint and permanent errors should be treated by permanently discarding
it. This removes the need for the implicit information that was passed in
the error number.