MXS-2622 Change MainWorker::get() to return pointer
In all cases where MainWorker::get() can be called an instance
should be available. However, it still feels wrong to guarantee
an instance when it in principle may be absent.

Further, in all other situations we have a pointer instead of a
reference to the current worker. Conceptually clearer if a
pointer is consistently used.