MXS-2622 Do not load users at Listener::listen()
If the users are loaded, then the caller must be a routing worker.
However, if we use Type::SHARED_TCP, the actual socket descriptor
must be assigned in the main worker. Moving the assignment there
will then cause a deadlock when the WorkerLocal sends a synchronous
message to all routing worker.

Now, since the loading of the users would anyway only be of value
for the calling routing worker and since that prevents the entire
startup procedure from being performed by the MainWorker, the best
approach is simply not to load the users.

What could be done here, would be to load all users by the MainWorker
and pre-populate all routing workers with that data.