Track current default database
The current default database is now exposed by the MXS_SESSION. This
removes the need to access protocol objects from filters and routers.

The client protocol module updates the current default database whenever a
change is detected. In most cases the default database change happens with
a COM_INIT_DB command but in some cases a plain SQL USE statement is
used. This necessitates parsing of the query but this shouldn't have a
visible effect on performance due to the coarse first-pass check with
strcasecmp.