Do thread-level initialization for MainWorker
The query classifier and modules weren't initialized for the mainworker
thread when they should be as it interacts with them. The mainworker
disables the query classification cache as it is a waste of memory for the
mainworker thread.

Removed the stray modules_thread_init and modules_thread_finish
declarations from maxbase.