MXS-1134 Make note of invalidation words
In order to perform invalidation, we need to have a mapping from
words (aka table names) to the nodes that should be invalidated
by those words.

To be able to efficiently remove a node from the previous mapping
we also need to store

(aka table names) that should cause a node to be invalidated.
We also need to store