Make initial slave connection count configurable
By making the number of slave servers that the session starts with
configurable, the total amount of connections to the cluster is reduced
while still retaining the possibility of dynamic load balancing between
all servers.

The load balancing is made possible by the slave selection code that takes
unconnected slaves into use if they are significantly better than any of
the current connected slaves.