MXS-2729 Provide means for calculating load of session
Session now has a function set_load() using which the load of
the session can be set. This function is intended to be called
whenever there is some activity on a DCB of the session.

The session retains the set load values for the last 30 seconds
and the load of the session is defined as the average of those
values.

When the load is set, we move the existing values as many steps
to the right (with excess values falling of) as there has been
seconds since the load was last set, and zero the values from
the beginning.

Likewise, when the load is queried, we first move and zero values
as appropriate. With this approach we do not need to update all
sessions every second, but have to do so only when there is
activity.

This will not allow us to strictly order the sessions according
to how much load they cause, but will nonetheless give a hint of
how much a particular session contributes to the load of a worker.