Fall back to pkill in run_ctrl_c.sh
The pkill executable is provided by the same package that provides ps
which is why it is more likely to be installed. Seems that the test VMs
don't always have killall installed.

Run the MaxScale process with the maxscale user instead of the root user
to prevent any new files from being owned by a different user.