Remove test_session_track unit test
The test contained raw binary data which makes it very hard to debug. With
the change in how the tracking is done, adapting it to the new mechanism
would be a lot of work. Instead, a system test which uses a live MariaDB
backend can be used to test it. This should also make sure that it
actually works with a live server.