Move transaction tracking into Session
This way the data in mxs::Reply can be used to track the state
changes.

Given that this feature offers a secondary mechanism for transaction
tracking and the primary mechanism is used regardless of whether the
secondary one is used, it is unlikely that this feature is a net positive
in practice. With this fact in mind, removing it would be reasonable.