Add C++ PCRE2 class
The class provides an interface somewhat similar to what std::regex
provides except that it uses the bundled PCRE2 library (far more stable
than the standard regex implementations). The intended use-case for this
is to replace the tee filter match/exclude functionality with use of
mxs::Regex.

Added test cases that check the code works as expected.